requirements.txt
rich
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
pymongo[srv]
python-dotenv
